+bug fixed: node features should not be allowed to be of different size for the full graph and sub graph, what produces different embeddings + DTYPE coherency issues fixed + Device coherency issues fixed + use of node mask vs node features fixed

+++ subgraph_matching_via_nn/utils/utils.py
@@ -12,8 +12,8 @@ from common.graph_utils import SubGraphGenerator
 from subgraph_matching_via_nn.utils.graph_utils import \
     node_indicator_from_edge_indicator
 
-TORCH_DTYPE = torch.float32
-NP_DTYPE = np.float32
+TORCH_DTYPE = torch.float64
+NP_DTYPE = np.float64
 
 
 def plot_indicator(w_list, labels):
